[type:refactor] refactor admin-listener.
