+++ shenyu-admin-listener/shenyu-admin-listener-consul/src/main/java/org/apache/shenyu/admin/listener/consul/ConsulDataChangedInit.java
